[enhance] split into `stream.rs`.
